Different approach to monkey patching
